Infrastructure stabilization: Fix namespace issues, add JobStatus enum, remove orphaned code
- Added Data.Entities project reference to Infrastructure
- Created JobStatus enum matching database Status column values
- Fixed BackgroundJob type references (was using incorrect plural BackgroundJobs)
- Fixed namespace imports (Data.Entities for DbContext, Data.Entities.Entities for entities)
- Removed IBackgroundJob interface reference (doesn't exist)
- Fixed SqlVector null check (nullable type, not .IsNull property)
- Deleted EventBusHostedService (referenced deleted OODA handlers)
- Commented out deleted service registrations in DependencyInjection.cs

Infrastructure now builds successfully with database-centric Jobs/Caching/Messaging subsystems intact.
API/Admin/Workers still reference deleted repository services - these need database-centric reimplementation.

+++ src/Hartonomous.Infrastructure/Jobs/Processors/CleanupJobProcessor.cs
@@ -3,6 +3,7 @@ using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using Hartonomous.Data;
 using Hartonomous.Data.Entities;
+using Hartonomous.Data.Entities.Entities;
 
 namespace Hartonomous.Infrastructure.Jobs.Processors;
 
@@ -184,7 +185,7 @@ public class CleanupJobProcessor : IJobProcessor<CleanupJobPayload>
     {
         // Delete old completed background jobs
         var jobs = await _context.Set<BackgroundJob>()
-            .Where(j => j.Status == JobStatus.Completed && j.CompletedAtUtc < cutoffDate)
+            .Where(j => j.Status == (int)JobStatus.Completed && j.CompletedAtUtc < cutoffDate)
             .OrderBy(j => j.CompletedAtUtc)
             .Take(batchSize)
             .ToListAsync(cancellationToken);
@@ -202,7 +203,7 @@ public class CleanupJobProcessor : IJobProcessor<CleanupJobPayload>
     {
         // Delete old failed/dead-lettered jobs
         var jobs = await _context.Set<BackgroundJob>()
-            .Where(j => (j.Status == JobStatus.Failed || j.Status == JobStatus.DeadLettered) &&
+            .Where(j => (j.Status == (int)JobStatus.Failed || j.Status == (int)JobStatus.DeadLettered) &&
                        j.CompletedAtUtc < cutoffDate)
             .OrderBy(j => j.CompletedAtUtc)
             .Take(batchSize)
